Refactor: CheckoutSheet as its own file. Next: CartView (we might re-use on main page and in final review CheckoutSheet
